address comments pt4
